feat(toggleRefinement): rename toggle -> toggleRefinement
