+++ dev/app/jquery/stories/toggleRefinement.stories.js
@@ -2,14 +2,14 @@ import { storiesOf } from 'dev-novel';
 import { wrapWithHitsAndJquery } from '../../utils/wrap-with-hits.js';
 import * as widgets from '../widgets/index.js';
 
-const stories = storiesOf('Toggle');
+const stories = storiesOf('ToggleRefinement');
 
 export default () => {
   stories.add(
     'default',
     wrapWithHitsAndJquery(containerNode => {
       window.search.addWidget(
-        widgets.toggle({
+        widgets.toggleRefinement({
           containerNode,
           attributeName: 'free_shipping',
           label: 'Free Shipping (toggle single value)',
