+++ src/connectors/index.js
@@ -37,7 +37,9 @@ export {
   default as connectStarRating,
 } from './star-rating/connectStarRating.js';
 export { default as connectStats } from './stats/connectStats.js';
-export { default as connectToggle } from './toggle/connectToggle.js';
+export {
+  default as connectToggleRefinement,
+} from './toggleRefinement/connectToggleRefinement.js';
 export {
   default as connectBreadcrumb,
 } from './breadcrumb/connectBreadcrumb.js';
