+++ src/widgets/index.js
@@ -42,7 +42,9 @@ export {
 export { default as sortBy } from '../widgets/sort-by/sort-by.js';
 export { default as starRating } from '../widgets/star-rating/star-rating.js';
 export { default as stats } from '../widgets/stats/stats.js';
-export { default as toggle } from '../widgets/toggle/toggle.js';
+export {
+  default as toggleRefinement,
+} from '../widgets/toggleRefinement/toggleRefinement.js';
 export { default as analytics } from '../widgets/analytics/analytics.js';
 export { default as breadcrumb } from '../widgets/breadcrumb/breadcrumb.js';
 export { default as menuSelect } from '../widgets/menu-select/menu-select.js';
