+++ src/widgets/toggleRefinement/defaultTemplates.js
